pyproject.toml
[project]
name = "odi-exercise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.0",
    "coverage>=7.12.0",
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "redis>=7.1.0",
]
